test: split component tests
